+++ annotate-bullets.lua
@@ -1,19 +1,39 @@
+-- This filter adds some annotations before and after bullet lists
+-- and also before and after individual bullets. It's a bit of a hack
+-- job because my skills aren't really there yet, but it works.
+-- 
+-- Some areas for improvement:
+-- *  Annotations before/after individual bullets are actually before/after
+--    Para and Plain blocks, because I couldn't work out how to iterate over
+--    bullets in a list. This means some things are incorrectly annotated
+--    as being bullets, which is only okay because only things annotated as
+--    being in bullet lists are kept in R. But it's not nice.
+--
+-- *  Annotations before/after bullet lists don't indicate the level of
+--    indentation because the BulletList element doesn't include this 
+--    information, and I couldn't work out how to infer it here. This 
+--    information is currently inferred from the data structure in R. 
+--
+-- *  Annotate code blocks so linebreaks aren't removed in R
+
+-- Not necessary but makes the output data more intuitive to read
 traverse = "topdown"
 
-local function annotate_bullet(el) 
+local function annotate_bullet(el, name) 
   return {
-    pandoc.Str("!begin-bullet!"),
+    pandoc.Str("!begin-" .. name .. "!"),
     el,
-    pandoc.Str("!end-bullet!")
+    pandoc.Str("!end-" .. name .. "!")
   }
 end
 
-Para = annotate_bullet
-Plain = annotate_bullet
+function Para(el)      return annotate_bullet(el, "bullet")    end
+function Plain(el)     return annotate_bullet(el, "bullet")    end
+function CodeBlock(el) return annotate_bullet(el, "codeblock") end
 
 local bullets_level = 0
 
-function BulletList (el) 
+function BulletList(el) 
   
   bullets_level = bullets_level + 1
   
@@ -26,3 +46,4 @@ function BulletList (el)
   return out
   
 end
+